requirements.txt
asgiref==3.8.1
asttokens==2.4.1
crispy-tailwind==1.0.3
decorator==5.1.1
Django==5.1
django-crispy-forms==2.3
executing==2.0.1
ipython==8.26.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
djangorestframework==3.15.2